requirements.txt
Flask
pymongo
pillow
transformers
requests
urllib3
werkzeug